protocol: refactoring
protocol/begin_request: switched u(16|32)_t to uint\1_t since ntoh[sl]
already does the byte fiddling for us, making byte unions redundant.

protocol/end_request: ditto protocol/begin_request

protocol/gs: changed definition from _(gs?|s)_##EX_T {}; typedef
             _(gs?|s_##EX_T _\1_##ALIAS_T to _gs?|s)_##ALIAS_T so that
	     use of forward declaration in protocol/types.hpp doesn't
	     emit diagnostics.

protocol/header: added default constructor; ditto protocol/begin_request

protocol/management_reply: shortened variable names

protocol/message: std::shared_ptr does not keep custom deleter in
                  template args, so Array_deleter moved to default
		  constructor

protocol/types: update of copyright header
                management_reply: switch template type int to uint8_t
                                  to better reflect in-class limits

+++ include/mosh/fcgi/protocol/types.hpp
@@ -1,7 +1,7 @@
 //! @file protocol/types.hpp Types used for FastCGI protocol
 /***************************************************************************
-* Copyright (C) 2007 Eddie                                                 *
-*               2011 m0shbear                                              *
+* Copyright (C) 2011 m0shbear                                              *
+*               2007 Eddie                                                 *
 *                                                                          *
 * This file is part of mosh-fcgi.                                          *
 *                                                                          *
@@ -118,7 +118,7 @@ namespace protocol {
 	 * @tparam padding_len Length of padding at the end of the record. This is needed to keep
 	 * the record size a multiple of chunkSize.
 	 */
-	template <int name_len, int value_len, int padding_len>
+	template <uint8_t name_len, uint8_t value_len, uint8_t padding_len>
 	struct Management_reply;
 
 	/*! @brief Data structure used to pass messages within the mosh-fcgi task management system
